backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.0
requests==2.31.0
openai>=1.40.0
Werkzeug==2.3.7
